+++ src/Service/Ssm/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added documentation in class's headers.
+- AWS enhancement: Added region "us-iso-east-1".
 
 ## 1.1.1